chore: Fix migration if table doesn't exist
The batch jobs is a legacy table that won't be present on all installs, make sure table exists before attempting to truncate.

+++ database/migrations/2025_05_07_202251_remove_batch_job_tables.php
@@ -1,5 +1,6 @@
 <?php
 
+use Filament\Schemas\Schema;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Support\Facades\DB;
 
@@ -11,8 +12,12 @@ return new class extends Migration
     public function up(): void
     {
         // Clenup the job tables since they are not used
-        DB::table('job_batches')->truncate();
-        DB::table('failed_jobs')->truncate();
+        if (Schema::hasTable('job_batches')) {
+            DB::table('job_batches')->truncate();
+        }
+        if (Schema::hasTable('failed_jobs')) {
+            DB::table('failed_jobs')->truncate();
+        }
     }
 
     /**
